move scene controls to a separate modal
